+++ prestaspot/includes/class-presta-spot-block.php
@@ -46,6 +46,7 @@ class Presta_Spot_Block
             'show_name' => (bool)($attributes['showName'] ?? true),
             'show_description' => (bool)($attributes['showDescription'] ?? true),
             'layout' => (string)($attributes['layout'] ?? ''),
+            'view_mode' => (string)($attributes['viewMode'] ?? ''),
         ));
     }
 }